+++ src/core/CoreGitDownloader.ts
@@ -3,27 +3,7 @@ import download from 'download-git-repo';
 import ora from 'ora';
 import chalk from 'chalk';
 import path from 'path';
-
-export type SupportedTemplate = 'vue2' | 'vue3';
-
-/**
- * 模板地址
- */
-export const templates: Record<SupportedTemplate, { url: string, description: string, downloadUrl: string, routerData: string }> = {
-  vue2: {
-    url: 'https://github.com/Tencent/tdesign-vue-starter.git',
-    description: 'TDesign Vue2 Starter',
-    downloadUrl: 'github.com.cnpmjs.org:Tencent/tdesign-vue-starter#main',
-    routerData: 'https://github.com/Tencent/tdesign-vue-starter/blob/develop/src/router/modules/components.ts'
-  },
-  vue3: {
-    url: 'https://github.com/Tencent/tdesign-vue-next-starter.git',
-    description: 'TDesign Vue3 Starter',
-    downloadUrl: 'github.com.cnpmjs.org:Tencent/tdesign-vue-next-starter#main',
-    routerData: ''  // TODO: add vue3 router config
-  }
-};
-
+import { SupportedTemplate, templates } from './CoreTemplate';
 export class CoreGitDownloader {
   /**
    * 下载工程目录，依据配置选择是否需要筛选不需要目录
@@ -54,6 +34,9 @@ export class CoreGitDownloader {
       fs.writeFileSync(packagePath, JSON.stringify(packageContent, null, 2), {
         encoding: 'utf8'
       });
+
+      // TODO: 写入后依据用户选择内容，清除部份内容
+
       console.log();
       console.log(chalk.green('👏  初始化项目完成！👏'));
       console.log();
